UI updates

+++ kqueen/blueprints/ui/views.py
@@ -6,6 +6,7 @@ from flask import abort
 from flask import Blueprint
 from flask import current_app
 from flask import flash
+from flask import jsonify
 from flask import redirect
 from flask import render_template
 from flask import request
@@ -14,9 +15,10 @@ from flask import url_for
 from kqueen.models import Cluster
 from kqueen.models import Provisioner
 from kqueen.wrappers import login_required
-from uuid import UUID
+from uuid import UUID, uuid4
 
 import logging
+import time
 
 logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
@@ -29,16 +31,22 @@ ui = Blueprint('ui', __name__, template_folder='templates')
 @login_required
 def index():
     username = current_app.config['USERNAME']
-
     clusters = []
+    healthy = 0
     for cluster in list(Cluster.list(return_objects=True).values()):
         data = cluster.get_dict()
-        # TODO: teach ORM to get related objects for us
-        prv = Provisioner.load(data['provisioner'])
-        data['provisioner'] = prv.name.value
-        clusters.append(data)
-    clustertable = ClusterTable(clusters)
+        if data and 'state' in data:
+            if 'Error' not in data['state']:
+                healthy = healthy + 1
 
+            # TODO: teach ORM to get related objects for us
+            try:
+                prv = Provisioner.load(data['provisioner'])
+                data['provisioner'] = prv.name.value
+            except:
+                pass
+            clusters.append(data)
+    clustertable = ClusterTable(clusters)
     provisioners = []
     for provisioner in list(Provisioner.list(return_objects=True).values()):
         data = provisioner.get_dict()
@@ -47,7 +55,12 @@ def index():
         provisioners.append(data)
     provisionertable = ProvisionerTable(provisioners)
 
-    return render_template('ui/index.html', username=username, clustertable=clustertable, provisionertable=provisionertable)
+    overview = {
+        'clusters': len(clusters),
+        'health': int((healthy / len(clusters)) * 100) if (healthy and clusters) else 100,
+        'username': username
+    }
+    return render_template('ui/index.html', overview=overview, clustertable=clustertable, provisionertable=provisionertable)
 
 
 @ui.route('/login', methods=['GET', 'POST'])
@@ -64,7 +77,7 @@ def login():
             next_url = request.form.get('next', '')
             if next_url:
                 return redirect(next_url)
-            return redirect(url_for('index'))
+            return redirect(url_for('.index'))
 
     return render_template('ui/login.html', error=error)
 
@@ -73,7 +86,7 @@ def login():
 def logout():
     session.pop('logged_in', None)
     flash('You were logged out', 'success')
-    return redirect(url_for('ui.index'))
+    return redirect(url_for('.index'))
 
 
 # catalog
@@ -121,9 +134,13 @@ def provisioner_delete(provisioner_id):
 
     # load object
     try:
+        used_provisioners = [c.provisioner.value for c in list(Cluster.list(return_objects=True).values())]
         obj = Provisioner.load(object_id)
-        obj.delete()
-        flash('Provisioner %s successfully deleted.' % obj.name, 'success')
+        if str(object_id) not in used_provisioners:
+            obj.delete()
+            flash('Provisioner %s successfully deleted.' % obj.name, 'success')
+        else:
+            flash('Provisioner %s is used by deployed cluster, cannot delete.' % obj.name, 'warning')
         return redirect('/')
     except NameError:
         abort(404)
@@ -138,6 +155,35 @@ def provisioner_delete(provisioner_id):
 def cluster_deploy():
     form = ClusterCreateForm()
     if form.validate_on_submit():
+        cluster_id = str(uuid4())
+        # Create DB object
+        try:
+            cluster = Cluster(
+                id=cluster_id,
+                name=form.name.data,
+                state='Deploying',
+                provisioner=form.provisioner.data,
+                kubeconfig={},
+            )
+            cluster.save()
+        except Exception as e:
+            flash('Could not create cluster %s.' % form.name.data, 'danger')
+            logging.error('Creating cluster %s failed with following reason: %s' % (form.name.data, repr(e)))
+            return redirect('/')
+
+        # Actually provision cluster
+        res = False
+        try:
+            prv = Provisioner.load(form.provisioner.data).engine_cls()
+            res = prv.provision(cluster_id)
+        except Exception as e:
+            flash('Could not create cluster %s.' % form.name.data, 'danger')
+            logging.error('Creating cluster %s failed with following reason: %s' % (form.name.data, repr(e)))
+            return redirect('/')
+        if res:
+            flash('Provisioning of cluster %s is in progress.' % form.name.data, 'success')
+        else:
+            flash('Could not create cluster %s.' % form.name.data, 'danger')
         return redirect('/')
     return render_template('ui/cluster_deploy.html', form=form)
 
@@ -153,13 +199,28 @@ def cluster_detail(cluster_id):
     # load object
     try:
         obj = Cluster.load(object_id)
+        obj.get_state()
     except NameError:
         abort(404)
 
+    # load information about clusters
+    try:
+        cluster = obj.get_dict()
+    except:
+        cluster = None
+        flash('Unable to load cluster', 'danger')
+
+    status = {}
+    if obj.get_state() == 'OK':
+        try:
+            status = obj.status()
+        except:
+            flash('Unable to get information about cluster', 'danger')
+
     return render_template(
         'ui/cluster_detail.html',
-        cluster=obj.get_dict(),
-        status=obj.status(),
+        cluster=cluster,
+        status=status,
     )
 
 
@@ -168,3 +229,38 @@ def cluster_detail(cluster_id):
 def cluster_delete(cluster_id):
     # TODO: actually deprovision cluster
     return redirect('/')
+
+
+@ui.route('/clusters/<cluster_id>/deployment-status')
+@login_required
+def cluster_deployment_status(cluster_id):
+    try:
+        object_id = UUID(cluster_id, version=5)
+    except ValueError:
+        logging.debug('%s not valid UUID' % cluster_id)
+        abort(404)
+
+    # load object
+    try:
+        obj = Cluster.load(object_id)
+    except NameError:
+        logging.debug('Cluster with UUID %s not found' % cluster_id)
+        abort(404)
+
+    res = 0
+    progress = 1
+    result = 'UNKNOWN'
+    try:
+        prv = obj.get_provisioner()
+        if prv:
+            data = prv.engine_cls().get(cluster_id)
+            result = data['state']
+            if data['state'] == 'Deploying':
+                progress = int((((time.time() * 1000) - data['build_timestamp']) / data['build_estimated_duration']) * 100)
+                if progress > 99:
+                    progress = 99
+            else:
+                progress = 100
+    except:
+        res = 1
+    return jsonify({'response': res, 'progress': progress, 'result': result})
